requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
pyperclip==1.8.2
pandas==2.2.1
google-generativeai==0.8.4
openpyxl==3.1.5
python-dotenv==1.1.0
Crawl4AI==0.5.0
playwright==1.51.0